tests/sctp: allow both ENOSPC and EINPROGRESS for CIPSO tag space errors
If the NetLabel subsystem in the kernel is configured to apply CIPSO
packet labels directly to the packets instead of the sockets, the
kernel will return EINPROGRESS on a failed connect(2) instead of
ENOSPC.  This is due to differences in how the CIPSO labels are
applied to network traffic in the two cases and how the error
conditions are able to be propagated back to userspace.

This patch allows both error codes and allows the SCTP tests to
function properly in both cases.

Signed-off-by: Paul Moore <paul@paul-moore.com>